refactor(manifest): split schema/loader from validator (Phase 10 follow-up)
Optional follow-up permitted by the plan's Phase 10 "Secondary
Benefit: Shrinking manifest.rs" section:

> "If after this extraction manifest.rs is still over the readability
>  target, it can be split in a follow-up, but the plan does not
>  mandate that split now."

Phase 10 (#139 + #144) extracted the env-policy logic (reserved list,
interpolation parsing, cycle detection) into `env_model`. The remaining
`manifest.rs` sat at 1473 LOC — schema structs + loader + a ~150 LOC
validator + ~1200 LOC of tests. This PR completes the shrink by
splitting schema/loader from the validator.

Split:

- `manifest/mod.rs` (522 LOC, was 1473) — public types
  (`AgentManifest`, `EnvVarDecl`, `HooksConfig`, `IdentityConfig`,
  `ClaudeMarketplaceConfig`, `ClaudeConfig`, `ManifestWarning`),
  `pub use` of `env_model::{JACKIN_*}` constants, `impl AgentManifest`
  with `load` and `display_name`, `mod validate;` declaration, and the
  19 loading/schema/`deny_unknown_fields`/display_name tests.
- `manifest/validate.rs` (962 LOC, new) — `impl AgentManifest` with
  the `validate`, `validate_env_ref`, `validate_env_interpolation`
  methods, the `is_valid_env_var_name` helper, and the 31 validation
  tests (`validate_*` + `is_valid_env_var_name_*`).

Rust's split-impl-block pattern (same as Phase 4 used for `AppConfig`)
lets `pub fn validate` live in `validate.rs` while `pub fn load` and
`pub fn display_name` stay in `mod.rs` — one `AgentManifest`, two
files. External callers see no difference.

Verbatim moves throughout. No function body edits. No signature
changes. No new types.

Test Preservation Policy: all 50 tests migrated verbatim, co-located
by subject:

- mod.rs — 19 tests (loading/schema/display_name)
- validate.rs — 31 tests (validate + env-var-name helper)

Test count: 427 → 427 (unchanged). No test deleted. No test weakened.

Visibility changes (one bump, scoped to `manifest`):
- `is_valid_env_var_name`: private → `pub(super)` so the split-impl
  block in `validate.rs` can call it. No external visibility widening.

No public API change. All 11 external references (in `src/env_resolver.rs`,
`src/env_model.rs`, `src/repo.rs`, `src/instance/*.rs`) resolve unchanged.

After this PR, `src/manifest/mod.rs` at 522 LOC is within the plan's
"acceptable dense module" band (400–600 LOC). `validate.rs` at 962 LOC
is dominated by tests (~830 LOC of tests, ~130 LOC of validator code);
the lib body itself is comfortably under the readability target.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 427 tests run: 427 passed, 0 skipped (unchanged)
