+++ src/manifest/validate.rs
@@ -1,76 +1,8 @@
+use super::{AgentManifest, EnvVarDecl, ManifestWarning};
 use crate::env_model::extract_interpolation_refs;
-pub use crate::env_model::{
-    JACKIN_DIND_HOSTNAME_ENV_NAME, JACKIN_RUNTIME_ENV_NAME, JACKIN_RUNTIME_ENV_VALUE,
-};
-use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
-use std::path::Path;
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct AgentManifest {
-    pub dockerfile: String,
-    #[serde(default)]
-    pub identity: Option<IdentityConfig>,
-    pub claude: ClaudeConfig,
-    #[serde(default)]
-    pub hooks: Option<HooksConfig>,
-    #[serde(default)]
-    pub env: BTreeMap<String, EnvVarDecl>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct EnvVarDecl {
-    #[serde(rename = "default")]
-    pub default_value: Option<String>,
-    #[serde(default)]
-    pub interactive: bool,
-    #[serde(default)]
-    pub skippable: bool,
-    pub prompt: Option<String>,
-    #[serde(default)]
-    pub options: Vec<String>,
-    #[serde(default)]
-    pub depends_on: Vec<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct HooksConfig {
-    pub pre_launch: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct IdentityConfig {
-    pub name: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct ClaudeMarketplaceConfig {
-    pub source: String,
-    #[serde(default)]
-    pub sparse: Vec<String>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-#[serde(deny_unknown_fields)]
-pub struct ClaudeConfig {
-    #[serde(default)]
-    pub marketplaces: Vec<ClaudeMarketplaceConfig>,
-    #[serde(default)]
-    pub plugins: Vec<String>,
-}
-
-#[derive(Debug, Clone)]
-pub struct ManifestWarning {
-    pub message: String,
-}
 
 /// Check that an env var name contains only `[A-Za-z0-9_]` and doesn't start with a digit.
-fn is_valid_env_var_name(name: &str) -> bool {
+pub(super) fn is_valid_env_var_name(name: &str) -> bool {
     !name.is_empty()
         && name.is_ascii()
         && !name.as_bytes()[0].is_ascii_digit()
@@ -78,18 +10,6 @@ fn is_valid_env_var_name(name: &str) -> bool {
 }
 
 impl AgentManifest {
-    pub fn load(repo_dir: &Path) -> anyhow::Result<Self> {
-        let manifest_path = repo_dir.join("jackin.agent.toml");
-        let contents = std::fs::read_to_string(&manifest_path)?;
-        Ok(toml::from_str(&contents)?)
-    }
-
-    pub fn display_name(&self, fallback: &str) -> String {
-        self.identity
-            .as_ref()
-            .map_or_else(|| fallback.to_string(), |id| id.name.clone())
-    }
-
     pub fn validate(&self) -> anyhow::Result<Vec<ManifestWarning>> {
         let mut warnings = Vec::new();
 
@@ -244,437 +164,6 @@ mod tests {
         assert!(!is_valid_env_var_name("MY.VAR"));
         assert!(!is_valid_env_var_name("MY$VAR"));
         assert!(!is_valid_env_var_name("A}B"));
-    }
-
-    #[test]
-    fn loads_manifest_with_plugins() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = ["code-review@claude-plugins-official"]
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(manifest.dockerfile, "Dockerfile");
-        assert!(manifest.claude.marketplaces.is_empty());
-        assert_eq!(manifest.claude.plugins.len(), 1);
-        assert!(manifest.identity.is_none());
-    }
-
-    #[test]
-    fn loads_manifest_with_marketplaces_and_plugins() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = ["superpowers@superpowers-marketplace"]
-
-[[claude.marketplaces]]
-source = "obra/superpowers-marketplace"
-sparse = ["plugins", ".claude-plugin"]
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(
-            manifest.claude.plugins,
-            vec!["superpowers@superpowers-marketplace"]
-        );
-        assert_eq!(manifest.claude.marketplaces.len(), 1);
-        assert_eq!(
-            manifest.claude.marketplaces[0],
-            ClaudeMarketplaceConfig {
-                source: "obra/superpowers-marketplace".to_string(),
-                sparse: vec!["plugins".to_string(), ".claude-plugin".to_string()],
-            }
-        );
-    }
-
-    #[test]
-    fn loads_manifest_marketplace_without_sparse() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[[claude.marketplaces]]
-source = "jackin-project/jackin-marketplace"
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(manifest.claude.marketplaces.len(), 1);
-        assert_eq!(
-            manifest.claude.marketplaces[0],
-            ClaudeMarketplaceConfig {
-                source: "jackin-project/jackin-marketplace".to_string(),
-                sparse: vec![],
-            }
-        );
-        assert!(manifest.claude.plugins.is_empty());
-    }
-
-    #[test]
-    fn loads_manifest_without_plugins_defaults_to_empty() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-
-[[claude.marketplaces]]
-source = "obra/superpowers-marketplace"
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert!(manifest.claude.plugins.is_empty());
-        assert_eq!(manifest.claude.marketplaces.len(), 1);
-    }
-
-    #[test]
-    fn loads_manifest_with_identity() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[identity]
-name = "Agent Smith"
-
-[claude]
-plugins = []
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(manifest.identity.as_ref().unwrap().name, "Agent Smith");
-    }
-
-    #[test]
-    fn display_name_uses_identity_when_present() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[identity]
-name = "Agent Smith"
-
-[claude]
-plugins = []
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(manifest.display_name("agent-smith"), "Agent Smith");
-    }
-
-    #[test]
-    fn display_name_falls_back_to_class_name() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(manifest.display_name("agent-smith"), "agent-smith");
-    }
-
-    #[test]
-    fn rejects_unknown_top_level_field() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-unknown_field = true
-
-[claude]
-plugins = []
-"#,
-        )
-        .unwrap();
-
-        let error = AgentManifest::load(temp.path()).unwrap_err();
-
-        assert!(error.to_string().contains("unknown field"));
-    }
-
-    #[test]
-    fn rejects_unknown_claude_field() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-typo = "oops"
-"#,
-        )
-        .unwrap();
-
-        let error = AgentManifest::load(temp.path()).unwrap_err();
-
-        assert!(error.to_string().contains("unknown field"));
-    }
-
-    #[test]
-    fn rejects_unknown_identity_field() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[identity]
-name = "Smith"
-typo = true
-
-[claude]
-plugins = []
-"#,
-        )
-        .unwrap();
-
-        let error = AgentManifest::load(temp.path()).unwrap_err();
-
-        assert!(error.to_string().contains("unknown field"));
-    }
-
-    #[test]
-    fn loads_manifest_with_hooks() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[hooks]
-pre_launch = "hooks/pre-launch.sh"
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(
-            manifest.hooks.as_ref().unwrap().pre_launch.as_deref(),
-            Some("hooks/pre-launch.sh")
-        );
-    }
-
-    #[test]
-    fn loads_manifest_without_hooks() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert!(manifest.hooks.is_none());
-    }
-
-    #[test]
-    fn rejects_unknown_hooks_field() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[hooks]
-pre_launch = "hooks/pre-launch.sh"
-post_launch = "bad"
-"#,
-        )
-        .unwrap();
-
-        let error = AgentManifest::load(temp.path()).unwrap_err();
-
-        assert!(error.to_string().contains("unknown field"));
-    }
-
-    #[test]
-    fn loads_manifest_with_static_env() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[env.RUNTIME]
-default = "docker"
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert_eq!(manifest.env.len(), 1);
-        let var = &manifest.env["RUNTIME"];
-        assert_eq!(var.default_value.as_deref(), Some("docker"));
-        assert!(!var.interactive);
-    }
-
-    #[test]
-    fn loads_manifest_with_interactive_env() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[env.PROJECT]
-interactive = true
-prompt = "Select a project:"
-options = ["project1", "project2"]
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        let var = &manifest.env["PROJECT"];
-        assert!(var.interactive);
-        assert_eq!(var.prompt.as_deref(), Some("Select a project:"));
-        assert_eq!(var.options, vec!["project1", "project2"]);
-    }
-
-    #[test]
-    fn loads_manifest_with_env_depends_on() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[env.PROJECT]
-interactive = true
-prompt = "Select:"
-options = ["a", "b"]
-
-[env.BRANCH]
-interactive = true
-depends_on = ["env.PROJECT"]
-prompt = "Branch:"
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        let var = &manifest.env["BRANCH"];
-        assert_eq!(var.depends_on, vec!["env.PROJECT"]);
-    }
-
-    #[test]
-    fn loads_manifest_with_skippable_env() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[env.API_KEY]
-interactive = true
-skippable = true
-prompt = "API key (optional):"
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        let var = &manifest.env["API_KEY"];
-        assert!(var.skippable);
-    }
-
-    #[test]
-    fn loads_manifest_without_env() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-"#,
-        )
-        .unwrap();
-
-        let manifest = AgentManifest::load(temp.path()).unwrap();
-
-        assert!(manifest.env.is_empty());
-    }
-
-    #[test]
-    fn rejects_unknown_env_field() {
-        let temp = tempdir().unwrap();
-        std::fs::write(
-            temp.path().join("jackin.agent.toml"),
-            r#"dockerfile = "Dockerfile"
-
-[claude]
-plugins = []
-
-[env.FOO]
-default = "bar"
-typo = true
-"#,
-        )
-        .unwrap();
-
-        let error = AgentManifest::load(temp.path()).unwrap_err();
-
-        assert!(error.to_string().contains("unknown field"));
     }
 
     #[test]
